Refactorizar funciones de la biblioteca libft
Se refactorizaron las funciones ft_memcpy, ft_memmove, ft_atoi, ft_strdup, ft_bzero, ft_calloc, ft_strlcpy, ft_memchr, ft_memcmp y ft_strncmp de la biblioteca libft. Se actualizaron los tipos de los parámetros n y len a size_t para mayor consistencia y se incluyó el archivo de cabecera libft.h en cada función. Esto mejora la legibilidad y mantenibilidad del código.

+++ libft/ft_bzero.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lumartin <lumartin@student.42.fr>          +#+  +:+       +#+        */
+/*   By: lumartin <lumartin@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 17:48:47 by lumartin          #+#    #+#             */
-/*   Updated: 2024/09/17 15:28:20 by lumartin         ###   ########.fr       */
+/*   Updated: 2024/09/25 18:11:23 by lumartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
