+++ libft/ft_memmove.c
@@ -6,13 +6,15 @@
 /*   By: lumartin <lumartin@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 17:49:51 by lumartin          #+#    #+#             */
-/*   Updated: 2024/09/25 15:56:58 by lumartin         ###   ########.fr       */
+/*   Updated: 2024/09/25 18:17:31 by lumartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, unsigned int len)
+#include "libft.h"
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	unsigned int	i;
+	size_t	i;
 
 	if (!dst && !src)
 		return (0);
